+++ src/java/com/jaeksoft/searchlib/scheduler/ExecutionAbstract.java
@@ -112,7 +112,7 @@ public class ExecutionAbstract {
 		}
 	}
 
-	public boolean waitForStart(int secTimeOut) throws InterruptedException {
+	public boolean waitForStart(long secTimeOut) throws InterruptedException {
 		long timeOut = System.currentTimeMillis() + secTimeOut * 1000;
 		while (timeOut > System.currentTimeMillis()) {
 			if (getLastExecution() != null)
